+++ src/scripts/types.ts
@@ -1,0 +1,9 @@
+import type { Emulator } from './emulator';
+
+declare global {
+    interface Window {
+			theEmulator: Emulator | undefined;
+			JSBEEB_RESOURCES: Record<string, string>
+			JSBEEB_DISC?: string
+		}
+}
